flask app/requirements.txt
Flask
pandas
scikit-learn
xgboost
joblib
google-cloud-storage
